repaint: improve a delayed repainting mechanism
This mechanism inhibited calls to the function DISPATCH-REPAINT and collected
remembered REGION-UNION of the previous region and the current one per sheet.
When the context of inhibited repaint was left, all repaints were dispatched.

The motivation to introduce the macro WITH-INHIBITED-DISPATCH-REPAINT was caused
by an observation, that (SETF SHEET-REGION) and (SETF SHEET-TRANSFORMATION) may
result in a cascade of such changes and multiple repaints of the same region.
However, the event queue already compresses repaint events, and it is unlikely
that the event is read inbetween repaints caused by changing the sheet geometry.
In other words the mechanism was not really used in practice. The correct method
to inhibit is REPAINT-SHEET (not DISPATCH-REPAINT).

Moreover update-mirror-geometry and writers for the sheet transformation and the
sheet region cause a call to repaint-sheet (not dispatch-repaint) to ensure,
that the repaint is performed immedietely (unless explicitly inhibited). This is
important when we draw on a sheet without reading events (i.e in a display
function) and call SCROLL-EXTENT which requires a repaint.